arreglos

+++ src/main/java/ar/edu/unq/epersgeist/servicios/interfaces/DegreeResult.java
@@ -2,6 +2,6 @@ package ar.edu.unq.epersgeist.servicios.interfaces;
 
 import ar.edu.unq.epersgeist.modelo.ubicacion.Ubicacion;
 
-public record DegreeResult(Ubicacion ubicacion, Double degree) { }
+public record DegreeResult(Ubicacion ubicacion, Long degree) { }
